% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{Autor, Dorn, and Hanson (2013)}
\format{
A data frame with 1444 rows and 243 variables
}
\usage{
df
}
\description{
A dataset containing commuting-zone panel data for 1990-2000 from Autor,
Dorn, and Hanson (2013). This dataset is in "long" format with shares
stored in \code{ssaggregate::shares}
}
\keyword{datasets}
