% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_wide}
\alias{df_wide}
\title{Autor, Dorn, and Hanson (2013)}
\format{
A data frame with 1444 rows and 640 variables
}
\usage{
df_wide
}
\description{
A dataset containing industry observations
}
\keyword{datasets}
