% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{industries}
\alias{industries}
\title{Autor, Dorn, and Hanson (2013)}
\format{
A data frame with 1444 rows and 640 variables
}
\usage{
industries
}
\description{
A dataset containing commuting-zone panel data for 1990-2000 from Autor,
Dorn, and Hanson (2013). This dataset is in a "wide" format with industry
shares as column variables
}
\keyword{datasets}
