% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shares}
\alias{shares}
\title{Autor, Dorn, and Hanson (2013)}
\format{
A data frame with 573268 rows and 4 variables:
\describe{
\item{czone}{Commuting Zone code}
\item{year}{Year}
\item{sic87dd}{Industry code (SIC87)}
\item{ind_share}{Employment share of industry in commuting zone}
}
}
\usage{
shares
}
\description{
A dataset of industry shares in "long" format. Each row is a commuting zone,
year, sic code with the corresponding industrial share.
}
\keyword{datasets}
