% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shocks}
\alias{shocks}
\title{Autor, Dorn, and Hanson (2013)}
\format{
A data frame with 794 rows and 21 variables
}
\usage{
shocks
}
\description{
A dataset containing industry-year observations of shocks
}
\keyword{datasets}
